sharpe ratio conversion and division by zero
The Sharpe ratio was still giving the wrong number. If convertrate is True, everything works properly. But if convertrate is False, the risk free rate is in annual terms and the returns of the strategy are in daily terms.

I am setting the sharpe ratio to None if there are not enough days of returns to calculate the standard deviation.